wyy 音乐 迷你播放器 暂未实现

+++ music/src/main/java/com/mio/music/data/L.kt
@@ -3,5 +3,5 @@ package com.mio.music.data
 data class L(val br: Int = 0,
              val fid: Int = 0,
              val size: Int = 0,
-             val vd: Int = 0,
+             val vd: Float = 0f,
              val sr: Int = 0)
